requirements.txt
# Core Framework
streamlit==1.31.0

# LLM Integration (Ollama)
ollama==0.1.6
requests==2.31.0
huggingface_hub==0.23.4
openai

# Environment & Configuration
python-dotenv==1.0.0

# Data Handling
pandas==2.2.0
python-json-logger==2.0.7

# Additional Utilities
validators==0.22.0